Старт дз 7, перенес функции добавления и удаления чатов в ChatList из App
